requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytest==8.1.1
faker==24.2.0
pandas==2.2.1
